+++ packages/encodable/src/parsers/scale/applyRound.ts
@@ -2,7 +2,7 @@ import { interpolateRound } from 'd3-interpolate';
 import { ScalePoint, ScaleBand, ScaleSymLog } from 'd3-scale';
 import { Value } from '../../types';
 import { D3Scale, ContinuousD3Scale } from '../../types/scale/Scale';
-import { HasToString } from '../../types/Base';
+import { StringLike } from '../../types/internal/Base';
 import { ScaleConfig } from '../../types/scale/ScaleConfig';
 
 export default function applyRound<Output extends Value>(
@@ -12,8 +12,8 @@ export default function applyRound<Output extends Value>(
   if ('round' in config && config.round === true) {
     const roundableScale = scale as
       | Exclude<ContinuousD3Scale<number>, ScaleSymLog<number, number>>
-      | ScalePoint<HasToString>
-      | ScaleBand<HasToString>;
+      | ScalePoint<StringLike>
+      | ScaleBand<StringLike>;
     if ('round' in roundableScale) {
       roundableScale.round(config.round);
     } else {
