+++ packages/encodable/src/types/internal/Base.ts
@@ -5,7 +5,7 @@ export type Unarray<T> = T extends Array<infer U> ? U : T;
 export type MayBeArray<T> = T | T[];
 
 /** A value that has .toString() function */
-export type HasToString = { toString(): string };
+export type StringLike = { toString(): string };
 
 /** Make some fields that might have been optional become required fields */
 export type RequiredSome<T, RequiredFields extends keyof T> = {
